Improved documentation and build script style fix.
Adds more detailed documentation how to build on Windows.
Fixes incorrect information how to build on Mac OS X.
Removes the special section for BSD suffixes from build.sh because the suffixes are currently the same.
If we figure out it needs different suffixes, we can add that again with the correct suffixes.