Enables gcPreserveCode along with some GC adjustments.
GCPreserveCode causes SpiderMonkey to keep JIT code instead of garbage collecting it. We trigger collecting of JIT code manually with a regular shrinking GC to avoid memory bloat. This reduces the frequency of JIT code GC compared to the default SpiderMonkey settings and improves performance.
Check this forum post for an in-depth description of the effects of gcPreserveCode:
http://www.wildfiregames.com/forum/index.php?showtopic=18466&p=300323