Now builds on Windows with Visual Studio 2010 or later.
Some parts of the code are now a bit uglier because some C++11 features aren't implemented in VS2010 yet and can't be used. Especially these constructors that had to be written manually are quite ugly. I still think it's not a big problem because we can make it less ugly as soon as we decide to drop VS2010 support and it shouldn't be a pain to maintain either.