CLogger: Use cppformat instead of sys_vswprintf.
sys_vswprintf relies on platform-specific printf implementations, which vary widely between platforms (in handling of truncation, return values, use of %s/%S/%hs/%ls for mixing char and wchar_t strings, etc) and are therefore a pain.

Use cppformat's fmt::sprintf instead, which has very similar syntax to sprintf but is more C++ish and is portable.

Also, wchar_t is stupid, so use char* strings (which are expected to be UTF-8) in CLogger. This creates a bit of a pain with changing all callers to convert to char* strings, but that's their fault for not using UTF-8 already.

Refs #3011.


git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16182 3db68df2-c116-0410-a063-a993310a9797