For material uniforms, let the engine decide if they are blocks.
The idea is that material files always provide a name and a blockname for uniforms. If the engine decides to store the uniforms in a block (because the hardware/driver supports it), it appends "[0]" to the uniform name and uses the blockname to identify the uniform block. If the engine decides to use plain uniforms, it uses the name as is and ignores the blockname.