Make GL_ARB_shader_draw_parameters emulation workaround optional
This workaround requires additional space in each vertex buffer, so only enable it if needed. Now it's enabled if GL4 is used, but it would be even better if GL_ARB_shader_draw_parameters would be used if it's available. Also rename InstancingDataOffset to VBContentOffset (more accurate name) and fix a problem in the calculation of the maximum number of vertices that fit into the buffer.