First step in integrating the old and the new renderer into the same codebase
The GLSL shaders still have to be replaced manually when switching between the renderers and recomiping is needed because the USEGL4 flag is a compile-time constant. I've used templates instead of inheritance and virtual functions where possible and where it seemed sensible because a lot of function calls happen per frame. Performance measurements weren't conclusive, but it seems like cache efficiency got slightly worse, which countered the positive effects of inlining.