Add material "sealing".
The idea is to "seal" a material when it's completely loaded or when one or more changes are applied. Sealing updates the relevant data for shaders (uniform blocks and, in the future, handles for bindless textures). Doing these kinds of updates when a set of changes has been made instead of doing them after each change should be more efficient, although the calls to "seal" could probably still be put on better spots where more unnecessary updates could be avoided.