Improved material management
The new renderer needs to store material data on the GPU efficiently. Storing material data directly with the model is bad because it uses more memory and requires more updates of the data on the GPU. This change moves the store for materials to a central place in the MaterialManager and implements an interface which makes sure that identical materials are only stored once. In addition, data from templates (red directly from the  material xml files) is now shared between multiple materials.

Fixes incorrect uniform handling for actor variations.